[C$] Finish the recent fix to the disambiguator.
It turns out that the recent fix was not complete: besides the main function
that disambiguates expressions, also some auxiliary functions that normally
build cast or binary expressions need to use a more general construction that
prevents an incorrect re-grouping of expressions during disambiguation. It also
turns out that such auxiliary functions were already attempting an adjustment,
but it was insufficient in general; given that these auxiliary functions now
call the recently added constructors/adjusters, these auxiliary functions can be
considerably simplified.

Thanks to Eric Smith for running the disambiguator on an example that revealed
the need to "finish" this fix to the disambiguator. Added a test with that
example.