[C$] Fix validator bug uncovered by testing.
When checking storage class specifiers of a function definition, we need to
adjust the test for file vs. block scope, given that, by the time the validation
function is called, we have already pushed a block scope for the body. We do
that by passing a flag saying whether the storage class specifiers are the ones
of a function definition or not.