[AleoBFT stake] Improve model of certificate creation.
Limit checks performed by endorsers to the components of certificates that are
also components of the proposals, while an author can also check the set of
endorsing addresses. Although we model certificate creation as an atomic event,
it is still important to model the implicit exchange of proposals and
endorsements accurately.

The needed checks on whole certificates are performed by validators when they
receive certificates -- more precisely, upon storing them into their DAGs.