[AleoBFT stake] Avoid some checks when storing certificates.
Under fault tolerance conditions, there is no need for a validator to check the
quorum requirements on references to previous certificates when storing a
certificate into the DAG. It suffices to check (as it already does) that the
signers of the certificate form a quorum, because then, because of fault
tolerance, there must be at least one correct validator who signed the
certificate, and who checked the quorum requirement on the certificate's
predecessors.

This change makes the proofs a bit more complicated, because now the
`previous-quorum` invariant becomes inter-dependent with the other invariants
proved by simultaneous induction. But all the adapted proofs work, of course.