Made allocate-fixnum-range in GCL 2.7.0 and beyond.  Eliminated warnings when building ACL2 with GCL 2.7.0.
Quoting a comment in (defxdoc note-8-7 ...):

; The ACL2 function allocate-fixnum-range is now a no-op for GCL versions 2.7.0
; and later.  Thanks to Camm Maguire for the suggestion, and the explanation
; that its effect would generally be only to waste space in GCL 2.7.0.

Thanks to Camm Maguire for helping to eliminate warnings when building
ACL2 in (a pre-release of) GCL 2.7.0.

Passed fresh "make regression" for GCL 2.7.0 and "make
regression-fresh" for SBCL.  An issue with books/demos/fp.lisp
regarding older GCL versions will be addressed in another commit
before long.