[FTY] Put some code in logic mode.
The recognizers, constructors, and deconstructors for the FTY aggregates are now
in guard-verified logic mode, which enables other tools that use them to be in
guard-verified logic mode, so that there is at least some static type checking.

The added program-mode declarations are a bit verbose, but let us see more
clearly what still needs to be put into guard-verified logic mode, which can be
done a little at a time. (An earlier attempt I made to put the whole file in
guard-verified logic mode, adding a few guards and theorems, causes a failure
elsewhere related to the inclusion of uncertified files in a string library
file.)

I also had to adapt some proofs in an smtlink file: see added comments to the
file.