[X86ISA] Remove unneeded hints.
This removes the :hints-l and :hints-t arguments supplied to defthm-unsigned-byte-p and defthm-signed-byte-p.  These should never be needed for calls that don't change the hyps of the corollaries, because the macro puts in hints that should always work.  And the manually-written hints were observed to be slowing down a proof due to not restricting the theory like the automatic hints.