x86isa: Introduce `testgen` a test generator for asmtest
`testgen` uses XED to iterate over x86 instructions and generate tests
for them. It has some restrictions on instructions it supports. The big
ones are it only works with instructions that operate on flags, gprs,
xmm registers, and immediates. Beyond introducing `testgen`, this also
introduces some scripts to make it easy to run all snippets with random
inputs.

At the moment there are a few issues:
  - XED errors when trying to encode some instructions. I don't know why
    exactly that is.
  - `asmtest` crashes with various errors when running some tests.
  - Some instructions fail tests because some flags are undefined. We
    handle undefined flags correctly but flags which are undefined
    only sometimes can't easily be dealt with in an automated fashion.

That being said, these issues are handled gracefully and I'm comfortable
having this merged into the community books.