Fixed bug that prevented use of acl2-doc.
A :DOC topic with name
X86ISA::5.50 UNCATEGORIZED "INSTRUCTIONS"
was not being handled properly by the acl2-doc browser.  More
precisely, the ACL2 code the fixes names with special characters was
not dealing with the double-quote character (").  Without going into
details: The fundamental problem is a mismatch between Common Lisp
syntax and Emacs Lisp syntax for symbols.  Thanks to Warren Hunt for
reporting this problem.