[C$] Add two ops on validation info for variables.
One is just for returning an irrelevant one (it should be really a feature of
fixtypes).

The other one is to check and coerce an expected value of that type to have that
type. This is useful in the absence of static guarantees about the info
annotating an AST, but we plan to add guards and rules to enable this to be
statically guaranteed, obviating the need for this operation eventually.