[AleoBFT] Add model of events.
Compared to previous versions, there are more kinds of events, because now we
are modeling the exchange of proposals, endorsements, and certificates, as
opposed to modeling certificate creation as an atomic event. Another (albeit
orthogonal) difference is that we do not require broadcasts to go to all
(correct) validators, but instead we incorporate into the broadcasting events an
indication of the validators to send the messages to.