[C$] Relax validator's checks on `__int128`.
Although the GCC documentation was somewhat ambiguous, we found practical code
suggesting that `signed` (or `__signed` or `__signed__`) can be used in
combination with `__int128`, although it doesn't change the meaning.

This commit also actually fixes the fact that, regardless of the signed
specifier, the validator failed to consider `unsigned`.

Thanks to Grant Jurgensen for finding and reporting that the validator was
rejecting practical code with `__signed__ __int128`.