[C$] Simplify `input-files`.
Eliminate the ability to generate `defconst`s for the intermediate stages, since
it does not seem useful after all. Also, in the programmatic interface, just
return the value for the final stage, i.e. the value assigned to the (only)
`defconst` in the event interface.

Thanks to Grant Jurgensen and Eric Smith for useful discussions on this.