[C$] Simplify `output-files`.
Eliminate the `:const-files` option to generate a `defconst` for the file set
resulting from pretty-printing the translation unit ensemble, when `:process` is
`:print`. Correspondingly, now the programmatic interface does not return the
file set. These changes mirror similar changes made to `input-files` a few
commits ago; the reason is that the removed features do not seem useful in
practice after all, after we have been using the tools for some time.

Thanks to Grant Jurgensen and Eric Smith for useful discussion on this topic.