[C$] Simplify `output-files`.
Eliminate the `:process` input. Now the macro always takes an (unambiguous)
translation unit ensemble as input, and pretty-prints to files in the file
system. The previous option to take a file set and write it to the file system
turned out to be not particularly useful, as we exercised these tools more. This
also makes the programmatic interface simpler, because it always takes a
translation unit ensemble, as opposed to either that or a file set.