[x86isa] Improve guard of rml08.
Previously its guard included both (canonical-address-p lin-addr) and (signed-byte 48 lin-addr), which mean the same thing.  I added :split-types so that the signed-byte-p claim is used only as a :type, not in the guard.  I then had to add (member-eq r-x '(:r :x)) to the guard so that it implies the :type.