[AleoBFT] Weaken a check in `propose` events.
We avoid checking explicitly that the authors of the previous certificates are
members of the committee at the previous round, because this should be an
invariant (as proved in other versions of the model). Indeed, we will make sure
that this is in fact the case in this version of the model and proofs.