[C$] Prove key properties of formalized subset.
The theorems show that the AST mapping is always defined (i.e. does not return
an error) when the predicates that characterize the formalized subset hold (by
'formalized subset' we mean the subset of C for which we have formal dynamic
semantics).

These theorems serve to justify the form of some upcoming theorems generated by
C-to-C transformations.