[C2C] Extend and refactor `simpadd0` proof generation.
Generalize the theorem generation function to expression to handle syntactically
equal expressions (for which the theorem only talks about the type), and to also
include an assertion about the type when the expressions are syntactically
unequal, besides asserting the equivalence of their executions. Also allow it to
omit the hypothesis that the old expression's execution does not return an
error, under the control of a flag, set when the expression cannot return an
error (under the hypotheses about the variables in scope).

Extend the theorem generated for `<var> + 0 --> <var>` transformation instances
with an assertion about the type of the expression, requiring the generation of
additional hints.