[Set-theory] Fixed terminology.  For f: A -> B, B is called a "codomain" of f (which I may sometimes call the range), but {f(x): x \in A} is called the "image" of f.  That is, the image is the set of values of f, and any superset of the image may be called the codomain of f.
I believe that the new terminology is correct.  Thanks to Eric Smith
for pointing out that my use of "codomain" previously (as the exact
image) was apparently non-standard.  As far as I can tell, "range" is
ambiguous; so mainly I'll stay away from that terminology, but it
seems to be OK to think of it as the codomain, so I plan to keep :ran
as a keyword for the zify macro.