[C$] Add "-std=c17" to C preprocessor calls
The Kestrel C tools use generally assume the C17 standard when it comes
to tests. Now, we explicitly pass the "-std=c17" flag to the
preprocessor by default.

This should fix the issue #1765 in which a test book failed to certify
when called with a newer version of GCC which defaulted to C23. Thanks
@copperteal for reporting the issue.