Made numerous tweaks to improve the hol-to-acl2 translation, in particular to handle goals, and added an example.
This continues to be joint work with Konrad Slind, who provided the translation of the new example.  That example, a simple polynomial evaluator, comes with a separate ACL2 formulation that we hope to use to prove an indicated goal, that evaluation distributes over addition.

Removed books/projects/hol-in-acl2/examples/thms1.lisp since the lone
theorem that was there, fst-comma, is now in
books/projects/hol-in-acl2/examples/ex1-thy.lisp (with a better
comment).