[doc] Improve instructions.
Add a step to commit the changes before merging in any remote changes.  This avoids having to talk about committing in 2 places (in case of a merge problem, and also before pushing).  We might as well always commit before merging in remote changes.  Note that neither the previous instructions nor this current version discuss what to do in case of a true merge conflict (incompatible changes to the same file).  We can add that later.