Work around SBCL Bug #2115955.
Certification was failing for books/demos/floating-point-book.lisp
when using ACL2 built on the latest SBCL, Version 2.5.6.  I've
reproduced the problem with a tiny, non-ACL2 example and submitted it
as the above SBCL bug.  The workaround here is a modification to ACL2
that avoids certain proclaiming for functions with df inputs or
outputs, for that version of SBCL or later (which I plan to restrict
to versions preceding a future SBCL version that fixes the bug).  As a
result, floating-point computations may be slower in those buggy SBCL
versions.