Created xdoc topics for the installation instructions, superseding installation/ directory, which has been deleted.  Made a few other documentation-related tweaks, including some small fixes.  Updated purity script to allow community to edit README.md.
Quoting :DOC note-8-7:

  The installation instructions have been moved into the manual; see
  installation.  Thanks to Eric Smith for suggesting this change,
  which will allow the ACL2 community (specifically, the acl2-books
  community) to improve those instructions.

My goal was to move the html-based installation instructions
(accessible from the ACL2 home page and from the installation/
directory of the ACL2 distribuion) into xdoc, with minimal
changes other than formatting (though I corrected errors and
tweaked wording when appropriate).

Now that those instructions are in xdoc, the acl2-books community is
welcome to improve them (just as they are already welcome to improve
the resst of the ACL2 documentation, in
books/system/doc/acl2-doc.lisp).

Quoting a comment from (defxdoc note-8-7 ...):

; As part of carrying out the item, "The installation instructions have been
; moved into the manual...", the installation/ directory has been removed and
; all-files.txt has been updated correspondingly.  From that directory,
; however, the following files have not been converted to xdoc, and instead,
; the first two have been placed in
; https://www.cs.utexas.edu/users/moore/acl2/contrib/ and the third has been
; deleted (as it is very likely no longer relevant): windows7.html,
; windows-gcl-jared.html, and installing-make.html.

NOTE: I might move the git commit hash (now in :DOC git-commit-hash)
back out of xdoc, since it might not be available at the time the
manual is built.  And I might change to a tag instead of a hash.