[axe/x86] Improve handling of PE executables.
Add an option to various tools to specify the number of existing stack words to assume disjoint from code and input regions.  For ELF64 and MACHO64, the default is 1 (for the saved return address), but for PE64 the default is 5 (1 word for the saved return address and 4 words providing space for registers to be saved to the stack).