[xdoc] Simplify XML entity machinery.
Now there are fewer things that must be kept in sync (4 things instead of 6), so it should be easier to add support for more XML entities.

I confirmed that the constant alists are the same (except one is reversed, like the others, because it is now generated, like the others -- and it has no duplicate keys).

I confirmed that the big string constant, *xml-entity-stuff*, is the same except for whitespace, which is now more regular since that string is generated.