[C] Update dynamic semantics to use statement values.
This is a conceptually simple change because the type of statement values is
currently isomorphic to the type of optional values. However, this paves the way
for uncommenting the `:none` case of statement value and to differentiate, in
the dynamic semantics, between statements terminating with a `return` without
expression and statement that terminate without a `return`.