[APT] Improve `isodata`.
Improve the detection of stobjs. The previous approach, namely checking whether
any of the transformed arguments or results are stobjs, was insufficient when no
transformed arguments or results are stobjs but the function includes some
untransformed stobjs: the stobj restrictions must still hold.

This commit makes use of the stobj flag recently added to `defmapping` (which
`definj`, `defsurj`, and `defiso` are based on). The flag says whether any of
the arguments or results of the domains are stobjs.