[C$] [C2C] Add and use block item annotations.
Now the validator annotates each block item that is a statement with the
validation table just before the block item.

The `simpadd0` transformation retrieves the annotation and passes it to the
specific transformation code, which for now does not actually use it, but this
will change in upcoming commits.

Adapt the rest of the code to handle/propoagate the added annotations.

Also improve some AST names related to the statement component of a block item
that is a statement.