Fixed inlining bug.  Made related but trivial source and :DOC modifications.  Fixed raw Lisp error for certain illegal function specs in defabsstobj events.  Fixed a comment in source function one-way-unify-p.  Fixed a buggy message from :pso.  Added a newline in a message from gc-verbose.
Quoting :DOC note-8-7:

  Fixed a bug that was preventing inlining, in some Lisps (SBCL and
  perhaps Allegro CL and CMUCL), of functions whose name ends in
  "$INLINE", such as those generated by [defun-inline].  Thanks to
  Grant Jurgensen for reporting this bug with a helpful example.

The problem was that in SBCL and probably the other aforementioned
Lisps, fmakunbound removes the ability to inline even after the
symbol-function is restored.  This issue was already accounted for in
source function install-defs-for-add-trip (though only for SBCL; now
that includes CMUCL and Allegro CL as well), but it needed to be
addressed in hcomp-restore-defs as well.

While I was at it, I added source functions inline-namep and
notinline-namep.  It should be easy to convert these to logic mode if
that becomes useful.  (But it didn't seem sufficiently important to
spend the time required to test and possibly debug, not to mention the
extra certification that would be required for books depending on
arithmetic-5.)

I also improved :DOC defun-inline slightly, to clarify that it doesn't
cause inlining for *1* functions.

Also quoting :DOC note-8-7:

  A [defabsstobj] event contains function specs that are each either a
  symbol or a list of the form (fn :kwd1 val1 ... :kwdn valn).  Each
  vali must be a symbol, but when this was not the case, a raw Lisp
  error could occur.  Now a clean error message is printed, and this
  requirement on the vali has been made explicit in the documentation
  for [defabsstobj].

Quoting comments in (defxdoc note-8-7 ...):

; Fixed a message printed by :pso that mistakenly referenced obsolete :DOC
; topic set-saved-output when there is no saved output to print.  The message
; now points to :DOC pso, and it also suggests turning on PROVE output.  Thanks
; to Warren Hunt for reporting the buggy message.

; When Lisp does not do anything with (gc-verbose t), the error message reports
; that.  But a newline was missing before returning NIL, as pointed out by Eric
; Smith.  That has been fixed.