[C$] Fix bug in parser.
This is another instance of the same bug fixed in a recent commit. Compound
literals may be parsed without going through the function to parse postfix
expressions (where the previous bug fix was made), and so we needed to make the
same fix in another place (namely, in the function to parse unary expressions).

Thanks to Grant Jurgensen for reporting a parser failure due to this bug.