[C2C] Slightly weaken some intermediate theorems.
Add hypotheses saying that there is at least one frame, for the theorems about
the execution of constructs in a function's body. The top-level theorems for
functions are not weakened; their proofs are just slightly extended to establish
the non-emptiness of the frame stack in the body (since the function is
executing, there is a frame for it).

Adapt the proofs of the intermediate theorems to propagate the new hypotheses
when composing sub-proofs for sub-constructs.

This added hypotheses are needed for upcoming extensions, whose supporting
theorems have those hypotheses.