[C$] Refactor and extend some ASTs.
We introduce an explicit AST fixtype for blocks, corresponding to compound
statements in the concrete syntax (grammar). This also adds support for GCC
label declarations in all three situations in which blocks are used: as compound
statements, as statement expressions, and as function bodies. (Label
declarations at the beginning of function bodies seem unnecessary, because
regular labels already have function scope, but GCC accepts them.)

This required a number of adaptation to the syntax tools as well as to the
transformation tools.