[C$] Add `alignof` unary expressions (GCC extension).
This allows the use of `_Alignof`, `__alignof`, and `__alignof__` in the same
way as `sizeof`, as an operator preceding a unary expression and forming a unary
expression. This is only supported if GCC extensions are enabled; otherwise,
only `_Alignof(<type-name>)` is allowed, which is a different kind of
expression.

The same potential ambiguities arise as with `sizeof`, and we handle them in the
same way.

The diff artificially shows the whole grammar as changes just because of the
recent addition of the git attributes for CR LF end of line as required by ABNF.
The only actual change is the addition of an alternative to `unary-expression`,
and the addition of some documentation comments mentioning this GCC extension.