[C$] Fix two bugs in the disambiguator.
The first bug, uncovered by Grant Jurgensen when running the tools on an
example, was resulting in a hard error in the disambiguator. The cause was an
unwarranted assumption made the disambiguator, namely that the declarator of a
function definition includes parameters that cause a new scope to be pushed into
the disambiguation table. But this is not the case in syntactically valid but
semantically invalid code, where semantic validity is checked by the validator,
after disambiguation. The disambiguator is generally written with that
awareness, but in this case there was an unwarranted assumption. This has been
fixed, in `dimb-fundef`, by checking that a new scope has been indeed pushed.

After making this fix, some previous test failed, uncovering a separate bug in
the disambiguator. This has to do with the fact that it is the innermost
function declarator, in the declarator of a function definition, that should
push the scope and not pop it. The added documentation provides an explanatory
example.